switching to openbox (and pcmanfm)
Sadly, xfce sets up the environment without my full control. This in effect
meant that I could not set $SSH_AUTH_SOCK to the value I wanted in my .profile,
since xfce chose to overwrite it after sourcing my profile. This caused me a lot
of pain and I want to avoid these kinds of errors out of my control in the
future.